test: move tests into their respective files and fix cross-platform issues
- Delete test_coverage_gaps.py catch-all, distribute tests into matching
  test files (test_inspector.py, test_formatter.py, test_scanner.py, etc.)
- Add test_scan_formatter.py for scan formatter tests
- Fix virtualenv detection test: create orig-prefix.txt in both Unix and
  Windows locations so test works on all platforms
- Fix chmod-based tests: use Path.read_text mock instead of chmod(0o000)
  which doesn't restrict reads on Windows
- Fix stdlib_module_names fallback test: use delattr instead of patch.object
  to handle Python 3.9 where the attribute doesn't exist
